+++ src/lib/util/trie.c
@@ -2779,7 +2779,7 @@ bool fr_trie_delete(fr_trie_t *ft, void const *data)
  *
  * @param[in] ft	to return node count for.
  */
-uint64_t fr_trie_num_elements(UNUSED fr_trie_t *ft)
+uint32_t fr_trie_num_elements(UNUSED fr_trie_t *ft)
 {
 	return 0;
 }
